Changes for NOC-377: Seed NocCodes2021 table from SSOT

+++ src/WorkBC.Data/Migrations/20240411194800_AddNocCodes2021.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.EntityFrameworkCore.Migrations;
+using System;
 
 #nullable disable
 
@@ -12,7 +13,7 @@ namespace WorkBC.Data.Migrations
                 name: "NocCodes2021",
                 columns: table => new
                 {
-                    Id = table.Column<short>(type: "smallint", nullable: false),
+                    Id = table.Column<short>(type: "int", nullable: false),
                     Code = table.Column<string>(type: "nvarchar(5)", maxLength: 5, nullable: true),
                     Title = table.Column<string>(type: "nvarchar(150)", maxLength: 150, nullable: true),
                     FrenchTitle = table.Column<string>(type: "nvarchar(250)", maxLength: 250, nullable: true)
